refactor: split util module into submodules
