+++ src/util/mod.rs
@@ -1,0 +1,16 @@
+pub mod paths;
+pub mod rebuild;
+pub mod timer;
+pub mod output;
+pub mod error;
+pub mod summary;
+
+pub use paths::*;
+pub use rebuild::*;
+pub use timer::*;
+pub use output::*;
+pub use error::*;
+pub use summary::*;
+
+#[cfg(test)]
+mod tests;
